src/5.Proje/requirements.txt
numpy>=1.26.0
scipy>=1.11.0
Pillow>=10.0.0
PyWavelets>=1.5.0
